requirements.txt
Django>=4.0
psycopg2-binary
gunicorn
django-cors-headers
channels
daphne
djangorestframework
django_admin_shell
django_redis
django-sesame
yt-dlp
ffmpeg
requests
mutagen
debugpy